pyproject.toml
[project]
name = "nonebot-plugin-summary"
version = "0.1.0"
description = "AI总结群聊消息，再也不用一直爬楼了！"
authors = [
    {name = "ChenXu233", email = "Woyerpa@outlook.com"},
]
dependencies = [
    "nonebot2>=2.3.3",
    "nonebot-plugin-chatrecorder>=0.6.2",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
    "nonebot-plugin-alconna>=0.53.0",
    "requests>=2.32.3",
    "httpx>=0.27.2",
    "nonebot-plugin-userinfo>=0.2.6",
]
requires-python = ">=3.9,<=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.ruff.lint]
ignore = ["E402"]